requirements.txt
Django==4.2.5 
Pillow==10.0.1 
django-ninja==0.22.2 
